modify: rename Properties -> Resolvers
